COMP: fix warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10898
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10898 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/GUI/vtkSlicerApplicationGUI.cxx
@@ -2754,7 +2754,7 @@ void vtkSlicerApplicationGUI::PackDual3DView ( )
       }
 
     // Pack the 3D viewers
-    for (int i=0; i < (this->Internals->ViewerWidgets.size() > 2 ? 2 : this->Internals->ViewerWidgets.size()); i++)
+    for (size_t i=0; i < (this->Internals->ViewerWidgets.size() > 2 ? 2 : this->Internals->ViewerWidgets.size()); i++)
       {
       vtkSlicerViewerWidget *viewer_widget 
         = this->Internals->ViewerWidgets[i];
